# Update EVP AES GCM and XTS benchmarks to reuse the key and cipher during the benchmark