--- a/tool/speed.cc
+++ b/tool/speed.cc
@@ -490,6 +490,20 @@ static bool SpeedAESGCM(const EVP_CIPHER *cipher, const std::string &name,
     return true;
   }
 
+  TimeResults results;
+  BM_NAMESPACE::UniquePtr<EVP_CIPHER_CTX> ctx(EVP_CIPHER_CTX_new());
+  const size_t key_len = EVP_CIPHER_key_length(cipher);
+  std::unique_ptr<uint8_t[]> key(new uint8_t[key_len]);
+  const size_t iv_len = EVP_CIPHER_iv_length(cipher);
+  std::unique_ptr<uint8_t[]> nonce(new uint8_t[iv_len]);
+  if (!TimeFunction(&results, [&]() -> bool {
+        return EVP_EncryptInit_ex(ctx.get(), cipher, NULL, key.get(), nonce.get());})) {
+    fprintf(stderr, "EVP_EncryptInit_ex failed.\n");
+    ERR_print_errors_fp(stderr);
+    return false;
+  }
+  results.Print(name +  " encrypt init");
+
   for (size_t chunk_byte_len : g_chunk_lengths) {
     if (!SpeedAESGCMChunk(cipher, name, chunk_byte_len, ad_len,
                           /*encrypt*/ true)) {
@@ -497,6 +511,13 @@ static bool SpeedAESGCM(const EVP_CIPHER *cipher, const std::string &name,
     }
   }
 
+  if (!TimeFunction(&results, [&]() -> bool {
+        return EVP_DecryptInit_ex(ctx.get(), cipher, NULL, key.get(), nonce.get());})) {
+    fprintf(stderr, "EVP_DecryptInit_ex failed.\n");
+    ERR_print_errors_fp(stderr);
+    return false;
+  }
+  results.Print(name +  " decrypt init");
   for (size_t chunk_byte_len : g_chunk_lengths) {
     if (!SpeedAESGCMChunk(cipher, name, chunk_byte_len, ad_len, false)) {
       return false;
@@ -606,13 +627,29 @@ static bool SpeedAEADChunk(const EVP_AEAD *aead, std::string name,
 }
 
 static bool SpeedAEAD(const EVP_AEAD *aead, const std::string &name,
-                      size_t ad_len, const std::string &selected) {
+                      size_t ad_len, const std::string &selected, enum evp_aead_direction_t dir) {
   if (!selected.empty() && name.find(selected) == std::string::npos) {
     return true;
   }
 
+  TimeResults results;
+  BM_NAMESPACE::ScopedEVP_AEAD_CTX ctx;
+  const size_t key_len = EVP_AEAD_key_length(aead);
+  std::unique_ptr<uint8_t[]> key(new uint8_t[key_len]);
+
+  if (!TimeFunction(&results, [&]() -> bool {
+    return EVP_AEAD_CTX_init_with_direction(
+        ctx.get(), aead, key.get(), key_len, EVP_AEAD_DEFAULT_TAG_LENGTH,
+        evp_aead_seal);
+  })) {
+    fprintf(stderr, "EVP_AEAD_CTX_init_with_direction failed.\n");
+    ERR_print_errors_fp(stderr);
+    return false;
+  }
+  results.Print(name + (dir == evp_aead_seal ? " seal " : " open ") + "init");
+
   for (size_t chunk_len : g_chunk_lengths) {
-    if (!SpeedAEADChunk(aead, name, chunk_len, ad_len, evp_aead_seal)) {
+    if (!SpeedAEADChunk(aead, name, chunk_len, ad_len, dir)) {
       return false;
     }
   }
@@ -621,17 +658,12 @@ static bool SpeedAEAD(const EVP_AEAD *aead, const std::string &name,
 
 static bool SpeedAEADOpen(const EVP_AEAD *aead, const std::string &name,
                           size_t ad_len, const std::string &selected) {
-  if (!selected.empty() && name.find(selected) == std::string::npos) {
-    return true;
-  }
+  return SpeedAEAD(aead, name, ad_len, selected, evp_aead_open);
+}
 
-  for (size_t chunk_len : g_chunk_lengths) {
-    if (!SpeedAEADChunk(aead, name, chunk_len, ad_len, evp_aead_open)) {
-      return false;
-    }
-  }
-
-  return true;
+static bool SpeedAEADSeal(const EVP_AEAD *aead, const std::string &name,
+                          size_t ad_len, const std::string &selected) {
+  return SpeedAEAD(aead, name, ad_len, selected, evp_aead_seal);
 }
 
 static bool SpeedSingleKEM(const std::string &name, int nid, const std::string &selected) {
@@ -802,20 +834,25 @@ static bool SpeedAES256XTS(const std::string &name, //const size_t in_len,
   });
 
   BM_NAMESPACE::UniquePtr<EVP_CIPHER_CTX> ctx(EVP_CIPHER_CTX_new());
-  // Call EVP_EncryptInit_ex once with the cipher and key, the benchmark loop will reuse both
-  if (!EVP_EncryptInit_ex(ctx.get(), cipher, nullptr, key.data(),
-                          iv.data())){
-    fprintf(stderr, "Failed to configure AES XTS encryption context.\n");
+  TimeResults results;
+
+  // Benchmark just EVP_EncryptInit_ex with the cipher and key, the encrypt benchmark loop will reuse both
+  if (!TimeFunction(&results, [&]() -> bool {
+        return EVP_EncryptInit_ex(ctx.get(), cipher, nullptr, key.data(),
+                                  iv.data());
+      })) {
+    fprintf(stderr, "EVP_EncryptInit_ex failed.\n");
     ERR_print_errors_fp(stderr);
     return false;
   }
+  results.Print(name + " encrypt init");
+
   // Benchmark initialisation and encryption
   for (size_t in_len : g_chunk_lengths) {
     in.resize(in_len);
     out.resize(in_len);
     std::fill(in.begin(), in.end(), 0x5a);
     int len;
-    TimeResults results;
     if (!TimeFunction(&results, [&]() -> bool {
           if (!EVP_EncryptInit_ex(ctx.get(), nullptr, nullptr, nullptr,
                                   iv.data()) ||
@@ -828,24 +865,27 @@ static bool SpeedAES256XTS(const std::string &name, //const size_t in_len,
       fprintf(stderr, "AES-256-XTS initialisation or encryption failed.\n");
       return false;
     }
-    results.PrintWithBytes(name + " init and encrypt",
+    results.PrintWithBytes(name + " encrypt",
                            in_len);
   }
 
   // Benchmark initialisation and decryption
-  // Call EVP_DecryptInit_ex once with the cipher and key, the benchmark loop will reuse both
-  if (!EVP_DecryptInit_ex(ctx.get(), cipher, nullptr, key.data(),
-                          iv.data())){
-    fprintf(stderr, "Failed to configure AES XTS encryption context.\n");
+  // Benchmark just EVP_DecryptInit_ex with the cipher and key, the decrypt benchmark loop will reuse both
+  if (!TimeFunction(&results, [&]() -> bool {
+        return EVP_DecryptInit_ex(ctx.get(), cipher, nullptr, key.data(),
+                                  iv.data());
+      })) {
+    fprintf(stderr, "EVP_DecryptInit_ex failed.\n");
     ERR_print_errors_fp(stderr);
     return false;
   }
+  results.Print(name + " decrypt init");
+
   for (size_t in_len : g_chunk_lengths) {
     in.resize(in_len);
     out.resize(in_len);
     std::fill(in.begin(), in.end(), 0x5a);
     int len;
-    TimeResults results;
     if (!TimeFunction(&results, [&]() -> bool {
           if (!EVP_DecryptInit_ex(ctx.get(), nullptr, nullptr, nullptr,
                                   iv.data()) ||
@@ -858,7 +898,7 @@ static bool SpeedAES256XTS(const std::string &name, //const size_t in_len,
       fprintf(stderr, "AES-256-XTS initialisation or decryption failed.\n");
       return false;
     }
-    results.PrintWithBytes(name + " init and decrypt",
+    results.PrintWithBytes(name + " decrypt",
                            in_len);
   }
 
@@ -2135,21 +2175,21 @@ bool Speed(const std::vector<std::string> &args) {
 #if !defined(OPENSSL_BENCHMARK)
      ||
      !SpeedKEM(selected) ||
-     !SpeedAEAD(EVP_aead_aes_128_gcm(), "AEAD-AES-128-GCM", kTLSADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_aes_128_gcm(), "AEAD-AES-128-GCM", kTLSADLen, selected) ||
      !SpeedAEADOpen(EVP_aead_aes_128_gcm(), "AEAD-AES-128-GCM", kTLSADLen, selected) ||
-     !SpeedAEAD(EVP_aead_aes_256_gcm(), "AEAD-AES-256-GCM", kTLSADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_aes_256_gcm(), "AEAD-AES-256-GCM", kTLSADLen, selected) ||
      !SpeedAEADOpen(EVP_aead_aes_256_gcm(), "AEAD-AES-256-GCM", kTLSADLen, selected) ||
-     !SpeedAEAD(EVP_aead_chacha20_poly1305(), "AEAD-ChaCha20-Poly1305", kTLSADLen, selected) ||
-     !SpeedAEAD(EVP_aead_des_ede3_cbc_sha1_tls(), "AEAD-DES-EDE3-CBC-SHA1", kLegacyADLen, selected) ||
-     !SpeedAEAD(EVP_aead_aes_128_cbc_sha1_tls(), "AEAD-AES-128-CBC-SHA1", kLegacyADLen, selected) ||
-     !SpeedAEAD(EVP_aead_aes_256_cbc_sha1_tls(), "AEAD-AES-256-CBC-SHA1", kLegacyADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_chacha20_poly1305(), "AEAD-ChaCha20-Poly1305", kTLSADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_des_ede3_cbc_sha1_tls(), "AEAD-DES-EDE3-CBC-SHA1",kLegacyADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_aes_128_cbc_sha1_tls(), "AEAD-AES-128-CBC-SHA1",kLegacyADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_aes_256_cbc_sha1_tls(), "AEAD-AES-256-CBC-SHA1",kLegacyADLen, selected) ||
      !SpeedAEADOpen(EVP_aead_aes_128_cbc_sha1_tls(), "AEAD-AES-128-CBC-SHA1", kLegacyADLen, selected) ||
      !SpeedAEADOpen(EVP_aead_aes_256_cbc_sha1_tls(), "AEAD-AES-256-CBC-SHA1", kLegacyADLen, selected) ||
-     !SpeedAEAD(EVP_aead_aes_128_gcm_siv(), "AEAD-AES-128-GCM-SIV", kTLSADLen, selected) ||
-     !SpeedAEAD(EVP_aead_aes_256_gcm_siv(), "AEAD-AES-256-GCM-SIV", kTLSADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_aes_128_gcm_siv(), "AEAD-AES-128-GCM-SIV",kTLSADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_aes_256_gcm_siv(), "AEAD-AES-256-GCM-SIV",kTLSADLen, selected) ||
      !SpeedAEADOpen(EVP_aead_aes_128_gcm_siv(), "AEAD-AES-128-GCM-SIV", kTLSADLen, selected) ||
      !SpeedAEADOpen(EVP_aead_aes_256_gcm_siv(), "AEAD-AES-256-GCM-SIV", kTLSADLen, selected) ||
-     !SpeedAEAD(EVP_aead_aes_128_ccm_bluetooth(), "AEAD-AES-128-CCM-Bluetooth", kTLSADLen, selected) ||
+     !SpeedAEADSeal(EVP_aead_aes_128_ccm_bluetooth(),"AEAD-AES-128-CCM-Bluetooth", kTLSADLen, selected) ||
      !Speed25519(selected) ||
      !SpeedSPAKE2(selected) ||
      !SpeedRSAKeyGen(selected) ||